+++ frontend/svelte/src/lib/components/canisters/SelectCyclesCanister.svelte
@@ -9,7 +9,8 @@
   import Input from "../ui/Input.svelte";
 
   export let amount: number | undefined = undefined;
-  export let icpToCyclesRatio: bigint | undefined = undefined;
+  export let icpToCyclesExchangeRate: bigint | undefined = undefined;
+  export let minimumCycles: number | undefined = undefined;
 
   let isChanging: "icp" | "tCycles" | undefined = undefined;
   let amountCycles: number | undefined;
@@ -19,19 +20,19 @@
 
   const setCycles = () =>
     (amountCycles =
-      amount !== undefined && icpToCyclesRatio !== undefined
+      amount !== undefined && icpToCyclesExchangeRate !== undefined
         ? convertIcpToTCycles({
             icpNumber: amount,
-            ratio: icpToCyclesRatio,
+            exchangeRate: icpToCyclesExchangeRate,
           })
         : undefined);
 
   const setAmount = () =>
     (amount =
-      amountCycles !== undefined && icpToCyclesRatio !== undefined
+      amountCycles !== undefined && icpToCyclesExchangeRate !== undefined
         ? convertTCyclesToIcpNumber({
             tCycles: amountCycles,
-            ratio: icpToCyclesRatio,
+            exchangeRate: icpToCyclesExchangeRate,
           })
         : undefined);
 
@@ -53,7 +54,10 @@
       amount: ICP.fromString(String(amount)),
     });
   };
-  // TODO: Add validations - https://dfinity.atlassian.net/browse/L2-644
+
+  let enoughCycles: boolean;
+  $: enoughCycles =
+    minimumCycles === undefined ? true : (amountCycles ?? 0) >= minimumCycles;
 </script>
 
 <div class="wizard-wrapper wrapper" data-tid="select-cycles-screen">
@@ -67,7 +71,7 @@
         bind:value={amount}
         on:focus={() => (isChanging = "icp")}
         on:blur={() => (isChanging = undefined)}
-        disabled={icpToCyclesRatio === undefined}
+        disabled={icpToCyclesExchangeRate === undefined}
       />
       <Input
         placeholderLabelKey="canisters.t_cycles"
@@ -77,7 +81,7 @@
         bind:value={amountCycles}
         on:focus={() => (isChanging = "tCycles")}
         on:blur={() => (isChanging = undefined)}
-        disabled={icpToCyclesRatio === undefined}
+        disabled={icpToCyclesExchangeRate === undefined}
       />
     </div>
     <slot />
@@ -86,7 +90,7 @@
     class="primary full-width"
     on:click={selectAccount}
     data-tid="select-cycles-button"
-    >{$i18n.canisters.review_cycles_purchase}</button
+    disabled={!enoughCycles}>{$i18n.canisters.review_cycles_purchase}</button
   >
 </div>
 
@@ -105,5 +109,6 @@
   .inputs {
     display: flex;
     gap: var(--padding-2x);
+    justify-content: center;
   }
 </style>