+++ frontend/svelte/src/lib/components/ui/PrincipalInput.svelte
@@ -8,7 +8,7 @@
   export let name: string;
   export let principal: Principal | undefined = undefined;
 
-  let address: string = "";
+  let address: string = principal?.toText() ?? "";
   $: principal = getPrincipalFromString(address);
   let showError: boolean = false;