+++ frontend/svelte/src/lib/stores/busy.store.ts
@@ -5,8 +5,11 @@ export type BusyStateInitiatorType =
   | "update-delay"
   | "vote"
   | "attach-canister"
+  | "detach-canister"
   | "create-canister"
   | "top-up-canister"
+  | "add-controller-canister"
+  | "remove-controller-canister"
   | "accounts"
   | "join-community-fund"
   | "split-neuron"
@@ -58,7 +61,7 @@ const initBusyStore = () => {
   };
 };
 
-const busyStore = initBusyStore();
+export const busyStore = initBusyStore();
 
 export const { startBusy, stopBusy } = busyStore;