+++ frontend/svelte/src/lib/types/neurons.errors.ts
@@ -1,9 +1,7 @@
 export class NotFoundError extends Error {}
 
-export class NotAuthorizedError extends Error {}
+export class NotAuthorizedNeuronError extends Error {}
 
 export class InvalidAmountError extends Error {}
-
-export class InsufficientAmountError extends Error {}
 
 export class CannotBeMerged extends Error {}