+++ frontend/svelte/src/lib/utils/utils.ts
@@ -52,6 +52,10 @@ export const stringifyJson = (
             return "Promise(...)";
           }
 
+          if (value instanceof ArrayBuffer) {
+            return new Uint8Array(value).toString();
+          }
+
           break;
         }
         case "bigint": {
@@ -126,8 +130,15 @@ export const bytesToHexString = (bytes: number[]): string =>
     ""
   );
 
-export const isNullOrUndefined = (value: unknown): boolean =>
-  value === undefined || value === null;
+/** Is null or undefined */
+export const isNullable = <T>(
+  argument: T | undefined | null
+): argument is undefined | null => argument === null || argument === undefined;
+
+/** Not null and not undefined */
+export const nonNullable = <T>(
+  argument: T | undefined | null
+): argument is NonNullable<T> => !isNullable(argument);
 
 export const mapPromises = async <T, R>(
   items: Array<T> | undefined,
@@ -139,3 +150,5 @@ export const mapPromises = async <T, R>(
 
   return Promise.all(items.map(async (item) => await fun(item)));
 };
+
+export const isArrayEmpty = <T>({ length }: T[]): boolean => length === 0;